Moved logic to service class
